Add handler authentication helper
